third_party/WebKit/LayoutTests/platform/android/http/tests/security/xss-eval-expected.txt

This page verifies that you can't use eval to subvert cross-domain checks.

If the test passes, you'll see a pass message below.

PASS: eval.call(frames[0], 'document').testExpando should be It's me! and is.
FAIL: childEval.call(frames[0], 'document').testExpando should be It's me too! but instead is EvalError.
PASS: childEvalCaller('document').testExpando should be SecurityError and is.
FAIL: childLocalEvalCaller('document').testExpando should be It's me too! but instead is EvalError.
